feat: refactoring the @CascadingUpdateShadowVariable listener logic
